refactor: ``members`` 엔티티 위치 변경
``members`` 엔티티 클래스의 위치를 ``global`` 패키지 하위가 되도록 변경하였습니다

Related to: #18
