+++ src/main/java/com/nextech/server/v1/domain/auth/service/impl/SignInServiceImpl.java
@@ -3,7 +3,7 @@ package com.nextech.server.v1.domain.auth.service.impl;
 import com.nextech.server.v1.domain.auth.dto.request.SignInRequest;
 import com.nextech.server.v1.domain.auth.service.PasswordValidationService;
 import com.nextech.server.v1.domain.auth.service.SignInService;
-import com.nextech.server.v1.domain.members.entity.Members;
+import com.nextech.server.v1.global.members.entity.Members;
 import com.nextech.server.v1.global.members.repository.MemberRepository;
 import com.nextech.server.v1.global.dto.response.TokenResponse;
 import com.nextech.server.v1.global.phonenumber.ConvertPhoneNumber;
