+++ src/main/java/com/nextech/server/v1/domain/members/service/impl/AllMembersInquiryServiceImpl.java
@@ -1,6 +1,6 @@
 package com.nextech.server.v1.domain.members.service.impl;
 
-import com.nextech.server.v1.domain.members.entity.Members;
+import com.nextech.server.v1.global.members.entity.Members;
 import com.nextech.server.v1.domain.members.service.AllMembersInquiryService;
 import com.nextech.server.v1.domain.members.dto.response.MembersInquiryResponse;
 import com.nextech.server.v1.global.members.repository.MemberRepository;
