+++ src/main/java/com/nextech/server/v1/domain/members/service/impl/DeleteProfilePictureServiceImpl.java
@@ -1,6 +1,6 @@
 package com.nextech.server.v1.domain.members.service.impl;
 
-import com.nextech.server.v1.domain.members.entity.Members;
+import com.nextech.server.v1.global.members.entity.Members;
 import com.nextech.server.v1.global.members.repository.MemberRepository;
 import com.nextech.server.v1.domain.members.service.DeleteProfilePictureService;
 import com.nextech.server.v1.global.aws.service.FileDeleteService;
