+++ src/main/java/com/nextech/server/v1/global/members/repository/MemberRepository.kt
@@ -1,6 +1,6 @@
 package com.nextech.server.v1.global.members.repository
 
-import com.nextech.server.v1.domain.members.entity.Members
+import com.nextech.server.v1.global.members.entity.Members
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
